Backend/requirements.txt
anyio==3.6.2
argcomplete==1.10.3
astroid==2.15.2
beautifulsoup4==4.8.2
chardet==3.0.4
click==8.1.3
compressed-rtf==1.0.6
dill==0.3.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
fastapi==0.95.0
h11==0.14.0
hazm==0.7.0
httptools==0.5.0
idna==3.4
IMAPClient==2.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
libwapiti==0.2.1
lxml==4.9.2
mccabe==0.7.0
nltk==3.3
olefile==0.46
pdfminer.six==20191110
Pillow==9.5.0
platformdirs==3.2.0
pycryptodome==3.17
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
starlette==0.26.1
strsimpy==0.2.1
textract==1.6.5
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
xlrd==1.2.0
XlsxWriter==3.0.9
